Change data type of conditionset and showconditionset in unlockable_t and extraemblem_t

+++ src/deh_soc.c
@@ -3127,9 +3127,9 @@ void readextraemblemdata(MYFILE *f, INT32 num)
 				deh_strlcpy(extraemblems[num-1].description, word2,
 					sizeof (extraemblems[num-1].description), va("Extra emblem %d: objective", num));
 			else if (fastcmp(word, "CONDITIONSET"))
-				extraemblems[num-1].conditionset = (UINT8)value;
+				extraemblems[num-1].conditionset = (UINT16)value;
 			else if (fastcmp(word, "SHOWCONDITIONSET"))
-				extraemblems[num-1].showconditionset = (UINT8)value;
+				extraemblems[num-1].showconditionset = (UINT16)value;
 			else
 			{
 				strupr(word2);
@@ -3215,9 +3215,9 @@ void readunlockable(MYFILE *f, INT32 num)
 				if (fastcmp(word, "HEIGHT"))
 					unlockables[num].height = (UINT16)i;
 				else if (fastcmp(word, "CONDITIONSET"))
-					unlockables[num].conditionset = (UINT8)i;
+					unlockables[num].conditionset = (UINT16)i;
 				else if (fastcmp(word, "SHOWCONDITIONSET"))
-					unlockables[num].showconditionset = (UINT8)i;
+					unlockables[num].showconditionset = (UINT16)i;
 				else if (fastcmp(word, "NOCECHO"))
 					unlockables[num].nocecho = (UINT8)(i || word2[0] == 'T' || word2[0] == 'Y');
 				else if (fastcmp(word, "NOCHECKLIST"))
